Fleshed out RRTCarrot Class
Continued development on RRTCarrot path planner. Added some helper
classes, RRTNode and RRTCarrotTree, which will be used by the planner to
implmenet RRT-connect. 

Needed to modify OccupancyGrid slightly due to a
couple of minor bugs and the need to add a way to get the size (volume)
of the occupancy grid to use in creating an estimate for the size of the
RRT trees. 

#16 #17