Fixed bug that was causing RRTCarrot to hang
A rounding error on the step vector that was essentially causing a
setp vector of 0,0,0 was causing RRTCarrot to hang since it would never
reach a goal or hit an obstacle. Fixed by both rounding at the vector
level rather than the point level, and by refusing to step if the vector
norm is smaller than step_size.

RRTCarrot will now relyably produce a path. However the path is a little
wonky... it seems to not necceisarily be connecting properly. But since
it now works as per definition, considering this commit as fixing #20