Radical modificaiton to the way LocalPlanner calculates distance to goal during tentacle selection #117
Previous local planner simply took the distance to goal from the last valid point and used that as the modifier. This works fine when the goal is beyond the arc of all tentacles. However, when the goal is insdide the arc
of the tentacles but obscured by an obstacle, it will cause issues because it relyes on the fundamental assumption that moving along the arc will get you 'further' than moving away from the goal in the euclidian sense.
This is generally true for highly curved arcs, but for the shallower ones it is not true at all; every step along the arc is roughly equavlent in length to the movement away from the goal.

Instead, Local Planner now keeps track of the distance between the current point on the tentacle and the goal. When this value stops decreasing we know we have found the minima, and can stop caluclating it for that tentacle.
This should ensure that the tentacle selection will be correctly biased to the 'longest tentacle that gets closest to goal' as was the original intention